pipenv: convert to Python module
Some Python packages (such as the to-be-written twitch-python package)
use pipenv as a Python module, not as an executable. If a package
depends on and imports pipenv, that package breaks if the package's
Python interpreter version differs from pipenv's Python interpreter
version.

Fix these problems by converting pipenv from a Python application to
a Python module. This will let me add twitch-python, which imports
pipenv, to Nixpkgs.

For compatibility, maintain nixpkgs.pipenv as a Python application.